scrim_stats_backend/requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
django-filter>=23.0
django-cors-headers>=4.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0

# Production dependencies (commented until needed)
# gunicorn>=21.2.0  # WSGI HTTP Server
# whitenoise>=6.5.0  # Static file serving
# psycopg2-binary>=2.9.7  # PostgreSQL adapter 